Bind recovery apply to reviewed evidence and fix token overflow
Resolve the confirmed Battery findings on the Phoebe side.

Recovery apply contract (ratified with Hugo): -apply is now bound to the
COMPLETE reviewed event set, not just its cardinality. Digest() hashes every
validated, de-duplicated event canonically (sorted, length-framed) instead of
the request-id set alone, dry-run prints it as event_set_sha256, and -apply
requires both -expected-count and -expected-digest, compared before a Valkey
client is constructed. Previously an artifact could be swapped for another of
the same count, and altering token counts or org attribution under the same
request ids did not even change the printed digest.

Recovery validation now rejects status codes outside the database CHECK range
100..599 (0 remains the NULL case). Such an event replayed into Valkey
successfully and then poisoned the drainer on INSERT.

fresh_input_tokens was an INTEGER generated column over unwidened operands, so
individually valid int32 counts could overflow the subtraction and make
PostgreSQL reject the INSERT, destroying the raw invalid evidence the ledger
exists to retain. It is now BIGINT over explicitly cast BIGINT operands.

The same unwidened subtraction was in rateWindowSQL's billable_prompt. That
projection runs over every event in the window before valid_usage filters, so
one malformed row failed the entire hour's rating with 22003 — a wider blast
radius than the rejected INSERT. Found by the new boundary integration test.

Wake readiness received the full comma-separated X-Saturn-Served-Model
allow-list rather than the model the request body selected. Readiness matches
/v1/models exactly, so a CSV value could never be found and every cold
multi-model request burned the whole wake budget. checkModelBinding now returns
the singular authorized model it matched, and the proxy wakes on that.

TestAbortRaceStress accepted any nonzero event count; it now asserts exactly N
events, N distinct trusted ids, and one event per client correlation id. Adds a
race-test Make target and CI step for the concurrency-sensitive proxy/waker
packages, which decide whether a billable attempt is emitted exactly once.

Tests: same-count/different-id and same-id/different-payload tamper cases for
the digest; status-code boundaries 0/99/100/599/600; live-Postgres int32
overflow boundary proving raw persistence, invalid-usage reporting and no rated
row; multi-model wake routes; and successful wake followed by a cold backend,
asserting the honest response with exactly one inference POST and no replay.

+++ internal/rating/store.go
@@ -310,7 +310,15 @@ resolved AS (
         ev.prompt_tokens,
         ev.cached_tokens,
         ev.completion_tokens,
-        GREATEST(ev.prompt_tokens - ev.cached_tokens, 0) AS billable_prompt,
+        -- Widen BEFORE subtracting. prompt_tokens and cached_tokens are INTEGER,
+        -- so individually valid engine evidence can overflow int32 on the
+        -- difference (e.g. 2147483647 - (-2147483648)). This projection runs over
+        -- EVERY event in the window before valid_usage filters anything, so an
+        -- int32 subtraction here fails the whole hour's rating with 22003 —
+        -- one malformed row would block all billing for that window, not just
+        -- its own. BIGINT operands keep the row computable; it is then excluded
+        -- from money by valid_usage and reported as an invalid-usage attempt.
+        GREATEST(ev.prompt_tokens::bigint - ev.cached_tokens::bigint, 0) AS billable_prompt,
         -- The C4 ladder: direct (a) wins; else derived base x premium (b) for
         -- fine-tune traffic; else the plain base rate (c) for a base-model endpoint.
         -- The rd and rpb join guards are mutually exclusive on the fine-tune marker,
